moving code for cache functions to utils file
