+++ internal/cortex/frontend/transport/handler.go
@@ -11,7 +11,6 @@ import (
 	"io"
 	"net/http"
 	"net/url"
-	"regexp"
 	"strconv"
 	"strings"
 	"syscall"
@@ -40,7 +39,6 @@ var (
 	errCanceled              = httpgrpc.Errorf(StatusClientClosedRequest, context.Canceled.Error())
 	errDeadlineExceeded      = httpgrpc.Errorf(http.StatusGatewayTimeout, context.DeadlineExceeded.Error())
 	errRequestEntityTooLarge = httpgrpc.Errorf(http.StatusRequestEntityTooLarge, "http: request body too large")
-	cacheableResponseCodes   = []int{http.StatusRequestTimeout, http.StatusGatewayTimeout, http.StatusBadRequest}
 )
 
 // HandlerConfig Config for a Handler.
@@ -59,8 +57,6 @@ type Handler struct {
 	log          log.Logger
 	roundTripper http.RoundTripper
 	lruCache     *lru.Cache
-	regex        *regexp.Regexp
-	errorExtract *regexp.Regexp
 
 	// Metrics.
 	querySeconds *prometheus.CounterVec
@@ -90,8 +86,6 @@ func NewHandler(cfg HandlerConfig, roundTripper http.RoundTripper, log log.Logge
 		log:          log,
 		roundTripper: roundTripper,
 		lruCache:     LruCache,
-		regex:        regexp.MustCompile(`[\s\n\t]+`),
-		errorExtract: regexp.MustCompile(`Code\((\d+)\)`),
 	}
 
 	if cfg.QueryStatsEnabled {
@@ -156,7 +150,7 @@ func (f *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	// Check if caching is enabled.
 	if f.lruCache != nil {
 		// Store query expression.
-		queryExpressionNormalized = f.regex.ReplaceAllString(r.URL.Query().Get("query"), " ")
+		queryExpressionNormalized = regex.ReplaceAllString(r.URL.Query().Get("query"), " ")
 
 		// Store query time range length.
 		queryExpressionRangeLength = getQueryRangeSeconds(r)
@@ -221,84 +215,6 @@ func (f *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	if f.cfg.QueryStatsEnabled {
 		f.reportQueryStats(r, queryString, queryResponseTime, stats)
 	}
-}
-
-func (f *Handler) updateFailedQueryCache(err error, queryExpressionNormalized string, queryExpressionRangeLength int, r *http.Request) {
-	// Extracting error code from error string.
-	codeExtract := f.errorExtract.FindStringSubmatch(err.Error())
-
-	// Checking if error code extracted successfully.
-	if codeExtract == nil || len(codeExtract) < 2 {
-		level.Error(util_log.WithContext(r.Context(), f.log)).Log(
-			"msg", "Error string regex conversion error",
-			"normalized_query", queryExpressionNormalized,
-			"range_seconds", queryExpressionRangeLength,
-			"error", err)
-		return
-	}
-
-	// Converting error code to int.
-	errCode, strConvError := strconv.Atoi(codeExtract[1])
-
-	// Checking if error code extracted properly from string.
-	if strConvError != nil {
-		level.Error(util_log.WithContext(r.Context(), f.log)).Log(
-			"msg", "String to int conversion error",
-			"normalized_query", queryExpressionNormalized,
-			"range_seconds", queryExpressionRangeLength,
-			"error", err)
-		return
-	}
-
-	// If error should be cached, store it in cache.
-	if !isCacheableError(errCode) {
-		level.Debug(util_log.WithContext(r.Context(), f.log)).Log(
-			"msg", "Query not cached due to non-cacheable error code",
-			"normalized_query", queryExpressionNormalized,
-			"range_seconds", queryExpressionRangeLength,
-			"error", err,
-		)
-		return
-	}
-
-	// Checks if queryExpression is already in cache, and updates time range length value to min of stored and new value.
-	if contains, _ := f.lruCache.ContainsOrAdd(queryExpressionNormalized, queryExpressionRangeLength); contains {
-		if oldValue, ok := f.lruCache.Get(queryExpressionNormalized); ok {
-			queryExpressionRangeLength = min(queryExpressionRangeLength, oldValue.(int))
-		}
-		f.lruCache.Add(queryExpressionNormalized, queryExpressionRangeLength)
-	}
-
-	level.Debug(util_log.WithContext(r.Context(), f.log)).Log(
-		"msg", "Cached a failed query",
-		"normalized_query", queryExpressionNormalized,
-		"range_seconds", queryExpressionRangeLength,
-		"error", err,
-	)
-
-}
-
-// isCacheableError Returns true if response code is in pre-defined cacheable errors list, else returns false.
-func isCacheableError(statusCode int) bool {
-	for _, errStatusCode := range cacheableResponseCodes {
-		if errStatusCode == statusCode {
-			return true
-		}
-	}
-	return false
-}
-
-// Time range length for queries, if either of "start" or "end" are not present, return 0.
-func getQueryRangeSeconds(r *http.Request) int {
-	start, err := strconv.Atoi(r.URL.Query().Get("start"))
-	if err != nil {
-		return 0
-	}
-	end, err := strconv.Atoi(r.URL.Query().Get("end"))
-	if err != nil {
-		return 0
-	}
-	return end - start
 }
 
 func (f *Handler) reportFailedQuery(r *http.Request, queryString url.Values, err error) {
